Make auto-complete as clojure-snippets play nice
There was a conflict between auto-complete and snippets because as
you type things like `def`, you want tab to produce the snippet, but
the auto-complete menu is up so it doesn't work. I updated the key
map in order to have <C-tab> work, which seems to get the job done fine.