[emacs] Support using sh-send-sexp in multiple shell buffers
Because I work with both clojurescript and clojure repls at the same
time, I wanted to adapt the solution I had found for sending sexp's
from a buffer to the shell buffer to allow for changing context.

I introduced a var, tk-shell-process-name, to allow me to set the
context, and I hooked into ivy to create
tk-choose-shell-function. This function looks for all of the shell
processes, and allows me to capture, and move to, a shell buffer of my
choice. Once this choice is made, C-c C-e will pipe sexp's to that
shell buffer, which allows easy context shifting.