[emacs] Further refine es-send-via-tmux
I ran into an issue try to pipe the following command to tmux:

```
cat ~/file.txt | awk '{ print $1 }'
```

It turned out the be because the ' and $ characters require different
shell interpretation.

After some tinkering, it became clear that surrounding with single
quotes, and escaping single quotes found in the actual command allows
the function to work as expected for my failing case.