[emacs] Add ability to run region in osx Terminal
These commands use applescript to allow selecting text that you want
to run in an existing osx Terminal session. I found a [useful Stack
Overflow Answer][how-to-launch-terminal-from-elisp] that I'd been
trying to figure out for a while (this functionality was really the
only reason I had learned tmux and tinkered with its configuration).

[how-to-launch-terminal-from-elisp]:https://emacs.stackexchange.com/a/22654